northwind-control-center/requirements.txt
Flask>=3.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
